Tmpbuild & long to int
Added tmpbuild to gitignore

Long is changed to int

+++ VagtPlan.ApiService/Controllers/DepartmentController.cs
@@ -27,7 +27,7 @@ namespace ApiService.Controllers
         }
 
         [HttpGet("get/{id}")]
-        public async Task<ActionResult<Department>> GetDepartment([FromRoute] long id)
+        public async Task<ActionResult<Department>> GetDepartment([FromRoute] int id)
         {
             var department = await _context.Departments.FindAsync(id);
 
@@ -54,7 +54,7 @@ namespace ApiService.Controllers
         }
 
         [HttpPut("edit/{id}")]
-        public async Task<IActionResult> EditDepartment([FromRoute] long id, [FromBody] DepartmentDTO departmentDTO)
+        public async Task<IActionResult> EditDepartment([FromRoute] int id, [FromBody] DepartmentDTO departmentDTO)
         {
             var department = await _context.Departments.FindAsync(id);
             if (department == null)
@@ -82,7 +82,7 @@ namespace ApiService.Controllers
         }
 
         [HttpDelete("delete/{id}")]
-        public async Task<IActionResult> DeleteDepartment([FromRoute] long id)
+        public async Task<IActionResult> DeleteDepartment([FromRoute] int id)
         {
             var department = await _context.Departments.FindAsync(id);
             if (department == null)
@@ -101,7 +101,7 @@ namespace ApiService.Controllers
             return NoContent();
         }
 
-        private bool DepartmentExists(long id)
+        private bool DepartmentExists(int id)
         {
             return _context.Departments.Any(e => e.Id == id);
         }
